+++ services/learn-card-network/brain-service/src/routes/boosts.ts
@@ -21,6 +21,8 @@ import {
     getBoostRecipientsSkipLimit,
 } from '@accesslayer/boost/relationships/read';
 
+import { deleteStorageForUri, setStorageForUri } from '@cache/storage';
+
 import {
     getBoostUri,
     isProfileBoostOwner,
@@ -395,6 +397,7 @@ export const boostsRouter = t.router({
             }
 
             await boost.save();
+            await setStorageForUri(uri, JSON.parse(boost.boost));
 
             return true;
         }),
@@ -436,7 +439,7 @@ export const boostsRouter = t.router({
                 });
             }
 
-            await deleteBoost(boost);
+            await Promise.all([deleteBoost(boost), deleteStorageForUri(uri)]);
 
             return true;
         }),