+++ services/learn-card-network/learn-cloud-service/src/helpers/learnCard.helpers.ts
@@ -1,3 +1,5 @@
+import { readFile } from 'node:fs/promises';
+
 import { generateLearnCard, LearnCard } from '@learncard/core';
 import { CryptoPlugin, CryptoPluginType } from '@learncard/crypto-plugin';
 import { DIDKitPlugin, DidMethod, getDidKitPlugin } from '@learncard/didkit-plugin';
@@ -8,7 +10,7 @@ import { CeramicPlugin, getCeramicPlugin } from '@learncard/ceramic-plugin';
 import { ExpirationPlugin, expirationPlugin } from '@learncard/expiration-plugin';
 import { LearnCardPlugin, getLearnCardPlugin } from '@learncard/learn-card-plugin';
 
-import didkit from '../didkit_wasm_bg.wasm';
+const didkit = readFile(require.resolve('@learncard/didkit-plugin/dist/didkit_wasm_bg.wasm'));
 
 export type EmptyLearnCard = LearnCard<
     [CryptoPluginType, DIDKitPlugin, ExpirationPlugin, VCTemplatePlugin, LearnCardPlugin]
@@ -35,7 +37,7 @@ export const getEmptyLearnCard = async (): Promise<EmptyLearnCard> => {
     if (!emptyLearnCard) {
         const cryptoLc = await (await generateLearnCard()).addPlugin(CryptoPlugin);
 
-        const didkitLc = await cryptoLc.addPlugin(await getDidKitPlugin(didkit));
+        const didkitLc = await cryptoLc.addPlugin(await getDidKitPlugin(await didkit));
 
         const expirationLc = await didkitLc.addPlugin(expirationPlugin(didkitLc));
 
@@ -53,7 +55,7 @@ export const getLearnCard = async (seed = process.env.LEARN_CLOUD_SEED): Promise
     if (!learnCards[seed]) {
         const cryptoLc = await (await generateLearnCard()).addPlugin(CryptoPlugin);
 
-        const didkitLc = await cryptoLc.addPlugin(await getDidKitPlugin(didkit));
+        const didkitLc = await cryptoLc.addPlugin(await getDidKitPlugin(await didkit));
 
         const didkeyLc = await didkitLc.addPlugin(
             await getDidKeyPlugin<DidMethod>(didkitLc, seed, 'key')