+++ examples/snap-chapi-example/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @learncard/snap-chapi-example
 
+## 1.0.87
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/react@2.6.23
+
+## 1.0.86
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/react@2.6.22
+
+## 1.0.85
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/react@2.6.21
+
 ## 1.0.84
 
 ### Patch Changes