+++ packages/learn-card-bridge-http/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @learncard/create-http-bridge
 
+## 1.1.56
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.0.18
+
+## 1.1.55
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.0.17
+
+## 1.1.54
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.0.16
+
 ## 1.1.53
 
 ### Patch Changes