+++ packages/learn-card-cli/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @learncard/cli
 
+## 3.1.18
+
+### Patch Changes
+
+-   Updated dependencies [[`81618f2`](https://github.com/learningeconomy/LearnCard/commit/81618f2bec3365924e4bd24d61dead4e77f462b3)]:
+    -   @learncard/learn-cloud-plugin@1.1.10
+    -   @learncard/init@1.0.18
+
+## 3.1.17
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/learn-cloud-plugin@1.1.9
+    -   @learncard/init@1.0.17
+
+## 3.1.16
+
+### Patch Changes
+
+-   Updated dependencies [[`92759e2`](https://github.com/learningeconomy/LearnCard/commit/92759e202017f1a9818ee781d216979e5fc64790)]:
+    -   @learncard/didkit-plugin@1.1.0
+    -   @learncard/init@1.0.16
+    -   @learncard/learn-cloud-plugin@1.1.8
+
 ## 3.1.15
 
 ### Patch Changes