+++ packages/react-learn-card/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @welibraryos/react-learn-card
 
+## 2.6.23
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.0.18
+
+## 2.6.22
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.0.17
+
+## 2.6.21
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.0.16
+
 ## 2.6.20
 
 ### Patch Changes